+++ Cargo.toml
@@ -5,6 +5,14 @@ edition = "2021"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
+[lib]
+name = "d1_filework"
+path = "src/lib.rs"
+
+[[bin]]
+name = "d1_load_trans"
+path = "src/main.rs"
+
 [dependencies]
 
 serde = "1.0.91"
